Added time condition for deleting content in temp_product
git-svn-id: https://192.168.1.50:9880/easyshop_svn/trunk@2259 2929c1f8-612a-3548-b275-2a983e7c7391

+++ application/scripts/clear_image_temp_folder.php
@@ -3,11 +3,17 @@
 $path = '../../assets/temp_product/*';
 $files = glob($path);
 foreach($files as $file){ 
-    if(is_dir($file)){
-        deleteDir($file);
-    }
-    else{
-         unlink($file);
+    $datecreated = date ("F d Y H:i:s.", filectime($file));
+    $now = date("Y-m-d H:i:s");
+    $interval = date_diff(new DateTime($now),new DateTime($datecreated));
+    //DELETE ALL CONTENT THAT IS AT LEAST 3 HOUR OLD
+    if(intval($interval->format('H'),10) > 3){
+        if(is_dir($file)){
+            deleteDir($file);
+        }
+        else{
+             unlink($file);
+        }
     }
 }
 echo 'Done. The directory '.$path.' has been emptied.';
